added use of AMILAB_SYSTEM_PROCESSOR + changes from int to size_t
git-svn-id: file:///home/karl/SVN2/amilab.svn/code/trunk@1719 52b0d089-85ac-4fb5-acf5-4e3172ea7aaf

+++ src/LanguageBase/Variables.hpp
@@ -47,7 +47,7 @@ class LanguageBase_EXPORT Variables{
 
   BasicVariable::ptr operator [](int i)
   {
-    if ((i>=0)&&(i<GetSize()))
+    if ((i>=0)&&(i<(int)GetSize()))
       return _vars[i];
     return BasicVariable::ptr();
   }
@@ -81,7 +81,7 @@ class LanguageBase_EXPORT Variables{
   void RemoveDefault( BasicVariable::ptr& defvarcontext );
 
   //--------------------------------------------
-  int GetSize()
+  size_t GetSize()
   {
     return _vars.size();
   }
